+++ src/pages/index.js
@@ -10,7 +10,7 @@ export default function Home() {
   useEffect(() => {
     if (!isLoading) {
       if (isAuthenticated) {
-        router.push('/ambient-archive');
+        router.push('/player');
       } else {
         router.push('/login');
       }
