+++ src/styles/pages/Player.module.css
@@ -530,3 +530,337 @@
     width: calc(100% - 20px);
   }
 }
+
+.playerContainer {
+    display: flex;
+    flex-direction: column;
+    min-height: 100vh;
+    background-color: #111111;
+  }
+  
+  .playerTopBar {
+    height: 60px;
+    background-color: #181818;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 0 20px;
+    border-bottom: 1px solid #333;
+  }
+  
+  .backButton {
+    background-color: transparent;
+    border: 1px solid #444;
+    color: #aaaaaa;
+    font-family: 'Archivo', sans-serif;
+    font-weight: 200;
+    font-size: 0.8rem;
+    padding: 8px 15px;
+    cursor: pointer;
+    transition: all 0.2s ease;
+    letter-spacing: 1px;
+    text-decoration: none;
+    display: inline-block;
+  }
+  
+  .backButton:hover {
+    border-color: #ffffff;
+    color: #ffffff;
+  }
+  
+  .sessionControls {
+    display: flex;
+    gap: 10px;
+  }
+  
+  .sessionButton {
+    background-color: transparent;
+    border: 1px solid #ffffff;
+    color: #ffffff;
+    font-family: 'Archivo', sans-serif;
+    font-weight: 200;
+    font-size: 0.8rem;
+    padding: 8px 15px;
+    cursor: pointer;
+    transition: all 0.2s ease;
+    letter-spacing: 1px;
+    text-transform: uppercase;
+  }
+  
+  .sessionButton:hover {
+    background-color: #ffffff;
+    color: #111111;
+  }
+  
+  /* Responsive styles */
+  @media (max-width: 768px) {
+    .playerTopBar {
+      padding: 0 15px;
+      height: 50px;
+    }
+    
+    .backButton, .sessionButton {
+      font-size: 0.7rem;
+      padding: 6px 12px;
+    }
+  }
+
+/* Additional styles for Player page */
+.hamburgerMenu {
+  width: 30px;
+  height: 24px;
+  background: transparent;
+  border: none;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  cursor: pointer;
+  padding: 0;
+}
+
+.hamburgerMenu span {
+  width: 100%;
+  height: 2px;
+  background-color: #ffffff;
+  display: block;
+  transition: all 0.3s ease;
+}
+
+/* Menu Overlay */
+.menuOverlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: rgba(0, 0, 0, 0.5);
+  opacity: 0;
+  visibility: hidden;
+  transition: all 0.3s ease;
+  z-index: 95;
+}
+
+.menuOverlay.menuOpen {
+  opacity: 1;
+  visibility: visible;
+}
+
+/* Side Menu */
+.sideMenu {
+  position: fixed;
+  top: 0;
+  left: 0;
+  height: 100vh;
+  width: 280px;
+  background-color: #181818;
+  color: #eaeaea;
+  display: flex;
+  flex-direction: column;
+  z-index: 100;
+  transform: translateX(-100%);
+  transition: transform 0.3s ease;
+  box-shadow: 5px 0 15px rgba(0, 0, 0, 0.2);
+}
+
+.sideMenu.menuOpen {
+  transform: translateX(0);
+}
+
+.sidebarHeader {
+  padding: 25px 20px;
+  border-bottom: 1px solid #333;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.sidebarTitle {
+  font-family: 'Archivo', sans-serif;
+  font-weight: 100;
+  font-size: 1.4rem;
+  letter-spacing: 4px;
+  margin: 0;
+  color: #ffffff;
+  text-transform: uppercase;
+}
+
+.closeMenuButton {
+  background: none;
+  border: none;
+  color: #aaa;
+  font-size: 1.5rem;
+  cursor: pointer;
+}
+
+.closeMenuButton:hover {
+  color: #fff;
+}
+
+.userInfo {
+  padding: 20px;
+  border-bottom: 1px solid #333;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.userInitials {
+  width: 60px;
+  height: 60px;
+  border-radius: 50%;
+  background-color: #2c2c2c;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-family: 'Archivo', sans-serif;
+  font-size: 1.5rem;
+  color: #fff;
+  margin-bottom: 15px;
+}
+
+.userName {
+  font-family: 'Archivo', sans-serif;
+  font-weight: 300;
+  font-size: 1rem;
+  color: #fff;
+  margin-bottom: 5px;
+  text-align: center;
+}
+
+.navigation {
+  padding: 20px 0;
+  flex-grow: 1;
+}
+
+.navLink {
+  display: block;
+  padding: 12px 20px;
+  font-family: 'Archivo', sans-serif;
+  font-weight: 200;
+  font-size: 0.9rem;
+  color: #aaa;
+  text-decoration: none;
+  transition: all 0.2s ease;
+  letter-spacing: 1px;
+  margin-bottom: 5px;
+  border-left: 3px solid transparent;
+}
+
+.navLink:hover {
+  background-color: #242424;
+  color: #fff;
+}
+
+.navLink.active {
+  background-color: #242424;
+  color: #fff;
+  border-left-color: #fff;
+}
+
+.sidebarFooter {
+  padding: 20px;
+  border-top: 1px solid #333;
+}
+
+.logoutButton {
+  width: 100%;
+  background-color: transparent;
+  border: 1px solid #444;
+  color: #aaa;
+  font-family: 'Archivo', sans-serif;
+  font-weight: 200;
+  font-size: 0.8rem;
+  padding: 10px 15px;
+  cursor: pointer;
+  transition: all 0.2s ease;
+  letter-spacing: 1px;
+  text-transform: uppercase;
+}
+
+.logoutButton:hover {
+  border-color: #fff;
+  color: #fff;
+}
+
+/* Update the existing play Collection button with more padding */
+.playButton {
+  background-color: transparent;
+  border: 1px solid #ffffff;
+  color: #ffffff;
+  font-family: 'Archivo', sans-serif;
+  font-weight: 200;
+  font-size: 0.8rem;
+  padding: 12px 20px; /* Increased padding here */
+  cursor: pointer;
+  transition: all 0.2s ease;
+  letter-spacing: 1px;
+  text-transform: uppercase;
+  margin-right: 10px;
+}
+
+/* Add simplified dropdown menu styles for Player page */
+.menuDropdown {
+  position: fixed;
+  top: 60px;
+  left: 20px;
+  width: 180px;
+  background-color: rgba(30, 30, 30, 0.9);
+  backdrop-filter: blur(10px);
+  border: 1px solid #333;
+  border-radius: 4px;
+  padding: 10px 0;
+  opacity: 0;
+  visibility: hidden;
+  transform: translateY(-10px);
+  transition: all 0.2s ease;
+  z-index: 100;
+  box-shadow: 0 5px 15px rgba(0, 0, 0, 0.3);
+}
+
+.menuDropdown.menuOpen {
+  opacity: 1;
+  visibility: visible;
+  transform: translateY(0);
+}
+
+.dropdownNav {
+  display: flex;
+  flex-direction: column;
+}
+
+.dropdownLink {
+  padding: 10px 15px;
+  font-family: 'Archivo', sans-serif;
+  font-weight: 200;
+  font-size: 0.9rem;
+  color: #ddd;
+  text-decoration: none;
+  transition: all 0.2s ease;
+  letter-spacing: 1px;
+}
+
+.dropdownLink:hover {
+  background-color: rgba(255, 255, 255, 0.1);
+  color: #fff;
+}
+
+.dropdownButton {
+  width: calc(100% - 20px);
+  margin: 5px 10px;
+  background-color: transparent;
+  border: 1px solid #444;
+  color: #ddd;
+  font-family: 'Archivo', sans-serif;
+  font-weight: 200;
+  font-size: 0.9rem;
+  padding: 8px 12px;
+  cursor: pointer;
+  transition: all 0.2s ease;
+  letter-spacing: 1px;
+  text-transform: uppercase;
+}
+
+.dropdownButton:hover {
+  border-color: #fff;
+  color: #fff;
+}
